// packages/serialized-query/test/data/people.d.ts
export declare const peopleDataset: () => {
    authenticated: {
        people: {
            a: {
                name: string;
                age: number;
                favoriteColor: string;
                createdAt: number;
                lastUpdated: number;
            };
            b: {
                name: string;
                age: number;
                favoriteColor: string;
                createdAt: number;
                lastUpdated: number;
            };
            c: {
                name: string;
                age: number;
                favoriteColor: string;
                createdAt: number;
                lastUpdated: number;
            };
            d: {
                name: string;
                age: number;
                favoriteColor: string;
                createdAt: number;
                lastUpdated: number;
            };
            e: {
                name: string;
                age: number;
                favoriteColor: string;
                createdAt: number;
                lastUpdated: number;
            };
        };
    };
};